Bug in BVH fixed, traversal now working!
Bounding boxes from triangle vertices was not being created correctly.
BVH traversal now working but I need to skip checking boxes when I have
found a triangle intersection that is closer than the intersection of
the box. Should be easily fixed by using the same intersection-object
and storing the distance to the box there.